// src/types.ts
export interface FlightProps {
  id: number;
  cp1: string;
  cp2: string;
  distance: string;
  frequency: string;
  identification: string;
  course: string;
  altitude: string;
  direction: string;
  velocity: string;
  temperature: string;
  th1: string;
  mh1: string;
  est: string;
  fuel: string;
  th2: string;
  mh2: string;
  act: string;
  ate: string;
  ata: string;
  rem1: string;

  tas: string;
  tc1: string;
  tc2: string;
  ch: string;
  leg: string;
  ete: string;
  eta: string;
  rem2: string;
}